Fixed compilation errors introduced by previous patch on MacOS.
Test Plan: make check all

Reviewers: stupaq-fb, pja

Reviewed By: pja

Subscribers: rocketspeed

Differential Revision: https://reviews.facebook.net/D34317

+++ src/proxy/proxy.cc
@@ -85,7 +85,7 @@ struct HostSessionMatrix {
  * @param num_workers The total number of workers.
  * @return A worker ID for the session, between 0 and num_workers - 1 inclusive.
  */
-static int WorkerForSession(int session, int num_workers) {
+static int WorkerForSession(int64_t session, int num_workers) {
   return session % num_workers;
 }
